Доработка bigPicture, объединение его с другим модулем за ненадобностью модульности.
